Sets up bootstrap and basic layout
