--- components/layout.js
@@ -1,9 +1,0 @@
-const Layout = (props) => (
-  <div>
-    <h1>Observatorio</h1>
-    { props.children }
-    <p>Footer</p>
-  </div>
-);
-
-export default Layout;
