+++ docker-compose.yml
@@ -10,7 +10,6 @@ services:
     ports:
       - "3000:3000"
     volumes:
-      - ./pages:/app/pages/
-      - ./components:/app/components/
+      - ./src:/app/src/
       - ./package.json:/app/package.json
       - ./package-lock.json:/app/package-lock.json
