+++ package.json
@@ -18,8 +18,9 @@
     "start": "next start -p $PORT"
   },
   "dependencies": {
-    "next": "^9.1.7",
-    "react": "^16.12.0",
-    "react-dom": "^16.12.0"
+    "bootstrap": "^4.4.1",
+    "next": "^9.3.0",
+    "react": "^16.13.0",
+    "react-dom": "^16.13.0"
   }
 }
