--- pages/about.js
@@ -1,10 +1,0 @@
-import Layout from '../components/layout';
-
-const About = () => ( 
-  <Layout>
-    <p>Este es el about del observatorio</p>
-  </Layout>
-);
-
-export default About;
-
