+++ src/pages/about.jsx
@@ -1,0 +1,6 @@
+const About = () => ( 
+  <p>Este es el about del observatorio</p>
+);
+
+export default About;
+
