+++ src/pages/index.jsx
@@ -1,0 +1,6 @@
+
+const Index = () => ( 
+  <p>Este es el root del observatorio</p>
+);
+
+export default Index;
